requirements-cloud.txt
# Optimized requirements for Cloud Run deployment
# Core FastAPI dependencies
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.9

# Scientific computing libraries (lighter versions)
numpy==1.21.6
opencv-python-headless==4.5.5.64  # Headless version for cloud
Pillow==9.5.0
tensorflow-cpu==2.11.0  # CPU-only version for Cloud Run
matplotlib==3.5.3
scikit-learn==1.0.2
scipy==1.7.3
pandas==1.3.5
protobuf==3.19.6

# File handling and utilities
aiofiles==23.1.0
setuptools==65.5.0
wheel==0.38.4

# Production server
gunicorn==20.1.0